requirements.txt
python-dotenv
kucoin-python
pandas